Move config setup to the InitializationProcessor
