+++ src/SonarScanner.MSBuild.PreProcessor/AnalysisConfigProcessing/Processors/IAnalysisConfigProcessor.cs
@@ -21,7 +21,7 @@
 using System.Collections.Generic;
 using SonarScanner.MSBuild.Common;
 
-namespace SonarScanner.MSBuild.PreProcessor.AnalysisConfigProcessing;
+namespace SonarScanner.MSBuild.PreProcessor.AnalysisConfigProcessing.Processors;
 
 public interface IAnalysisConfigProcessor
 {
